refactor: resolve vscode type errors in wakatime card render and remove redundant css (#3232)
* refactor: resolve vscode type errors in wakatime card render and remove redundant css

* dev
